requirements.txt
pandas~=2.2.2
apache-airflow-providers-microsoft-azure
azure-servicebus

Faker~=30.8.2